verificacion registro

+++ car-flex/MongoDB/publicacion_eliminar.php
@@ -1,8 +1,6 @@
 <?php
-
-use MongoDB\Operation\DeleteOne;
-
     include("Conexion_MongoDB.php");
+    $fotoPredeterminada='img\fotosPubli\Foto_predeterminada.webp';
 
     if(isset($_POST['_id_publi'])){
         $_id_Publi=$_POST["_id_publi"];
@@ -10,7 +8,9 @@ use MongoDB\Operation\DeleteOne;
         $eliminarPubli=$publicaciones->deleteOne(
             ['fotoPubli' => $_id_Publi]   
         );
-        unlink($_id_Publi);
+        if($_id_Publi!=$fotoPredeterminada){
+            unlink($_id_Publi);
+            }
     }
     header("location:perfil.php");
 ?>
